+++ src/libs/Avatars/CustomAvatarCatalog.ts
